eventFilter: Refactor (séparation de la class Search et script eventList.js)
